requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
dj-database-url==2.2.0
Django==4.2.13
django-ckeditor==6.7.1
django-heroku==0.3.1
django-js-asset==2.2.0
django-livereload-server==0.5.1
exceptiongroup==1.2.1
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
livereload==2.6.3
livereload-server==0.1.0
multidict==6.0.5
nodeenv==1.9.1
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
watchfiles==0.22.0
whitenoise==6.6.0
yarl==1.9.4
# Cloudinary pour le stockage externe des images
django-cloudinary-storage==0.3.0
cloudinary==1.36.0
